+++ python/iqm/qdmi/__init__.py
@@ -25,3 +25,16 @@ __all__ = ["IQM_QDMI_CMAKE_DIR", "IQM_QDMI_INCLUDE_DIR", "IQM_QDMI_LIBRARY_PATH"
 
 def __dir__() -> list[str]:
     return __all__
+
+
+# If pickling is not supported by Qiskit's DataBin container, patch it.
+# This shim is for compatibility with older Qiskit versions and can be removed once Qiskit >= 2.1.0 is required.
+try:
+    import pickle  # noqa: S403
+
+    from qiskit.primitives.containers.data_bin import DataBin
+
+    pickle.loads(pickle.dumps(DataBin()))  # noqa: S301
+except NotImplementedError:
+    # Bypass the immutable __setattr__ restriction during unpickling
+    DataBin.__setattr__ = object.__setattr__